eslint + flow config
