+++ .flowconfig
@@ -1,11 +1,14 @@
 [ignore]
+.*/node_modules/.*
 
 [include]
 
 [libs]
+flow-typed
 
 [lints]
 
 [options]
+all=true
 
 [strict]
