+++ package.json
@@ -1,69 +1,29 @@
 {
   "name": "maci",
-  "version": "0.1.0",
-  "private": true,
+  "version": "0.0.1a",
+  "description": "Minimal Anti-Collusion Infrastructure",
+  "repository": "git@github.com:kendricktan/maci.git",
+  "author": "Kendrick Tan <kendricktan0814@gmail.com>",
+  "license": "MIT",
+  "scripts": {
+    "start": "flow-node app/index.js",
+    "circuit:compile": "mkdir -p app/circuits && circom circuits/circuit.circom -o app/circuits/circuit.json",
+    "circuit:setup": "cd app/circuits && snarkjs setup",
+    "circuit:generateverifier": "mkdir -p contracts && cd app/circuits && snarkjs generateverifier && mv verifier.sol ../../contracts/Verifier.sol"
+  },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "circom": "^0.0.31",
-    "circomlib": "^0.0.14",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1",
-    "snarkjs": "^0.1.18",
+    "circomlib": "^0.0.16",
+    "express": "^4.17.1",
+    "flow-bin": "^0.107.0",
+    "snarkjs": "^0.1.19",
     "websnark": "^0.0.5"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "circuit:compile": "mkdir -p src/circuits && circom circuits/circuit.circom -o src/circuits/circuit.json",
-    "circuit:setup": "cd src/circuits && snarkjs setup",
-    "circuit:generateverifier": "mkdir -p contracts && cd src/circuits && snarkjs generateverifier && mv verifier.sol ../../contracts/Verifier.sol",
-    "gg": "node src/legacy/Add_New_Leaf/generate_circuit_input.js"
-  },
-  "eslintConfig": {
-    "extends": [
-      "standard",
-      "standard-react",
-      "plugin:flowtype/recommended"
-    ],
-    "parser": "babel-eslint",
-    "plugins": [
-      "flowtype"
-    ],
-    "settings": {
-      "flowtype": {
-        "onlyFilesWithFlowAnnotation": false
-      }
-    }
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-flow": "^6.23.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-config-standard-react": "^7.0.2",
-    "eslint-plugin-flowtype": "^3.11.1",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.14.2",
-    "eslint-plugin-standard": "^4.0.0",
-    "flow-bin": "^0.102.0",
-    "flow-remove-types": "^2.103.0",
-    "ganache-cli": "^6.4.4",
-    "truffle": "^5.0.26",
-    "truffle-hdwallet-provider": "^1.0.12"
+    "babel-eslint": "^10.0.3",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-config-standard-flow": "^1.0.1",
+    "eslint-plugin-flowtype": "^4.3.0"
   }
 }
